Removed env file that is no longer needed
The environment used by wsynphot should be logically same to starkit since it imports starkit for carrying some of its functionalities. The current env file of wsynphot is outdated in comparison to starkit. To keep it automatically up-to-date with the env file of starkit, Travis CI and user guide steps have been set up to directly access the env file from starkit. Thus there is no need of placing this outdated env file in wsynphot and to avoid further confusion, it is deleted.