Resolved YAMLLoadWarning due to None value in dict fetched
This change is irrespective of python version (2 or 3). The warning generates repeatedly when config file is accessed from `wsynphot/data/default_wsynphot_config.yml` as the None value is accessed for the key in parsed dictionary. It is resolved as pointed in by warning message, the correct use of yaml.load() function: https://msg.pyyaml.org/load